requirements.txt
fastapi
uvicorn
openai
pygame
gtts
python-dotenv
stripe
requests
